[mypyc] Refactor: move mypyc.prebuildvisitor under mypyc.irbuild
It's conceptually part of the IR transform.
