+++ mypyc/irbuild/builder.py
@@ -30,7 +30,7 @@ from mypy.visitor import ExpressionVisitor, StatementVisitor
 from mypy.util import split_target
 
 from mypyc.common import TEMP_ATTR_NAME, TOP_LEVEL_NAME
-from mypyc.prebuildvisitor import PreBuildVisitor
+from mypyc.irbuild.prebuildvisitor import PreBuildVisitor
 from mypyc.ir.ops import (
     BasicBlock, AssignmentTarget, AssignmentTargetRegister, AssignmentTargetIndex,
     AssignmentTargetAttr, AssignmentTargetTuple, Environment, LoadInt, Value,
