+++ mypyc/irbuild/main.py
@@ -30,7 +30,7 @@ from mypy.build import Graph
 from mypyc.errors import Errors
 from mypyc.options import CompilerOptions
 from mypyc.ir.module_ir import ModuleIR, ModuleIRs
-from mypyc.prebuildvisitor import PreBuildVisitor
+from mypyc.irbuild.prebuildvisitor import PreBuildVisitor
 from mypyc.irbuild.vtable import compute_vtable
 from mypyc.irbuild.prepare import build_type_map
 from mypyc.irbuild.builder import IRBuilder
